requirements.txt
beautifulsoup4
bleach
bleach[css]
dkimpy
Flask
Flask-Caching
Flask-Compress
Flask-Limiter
Flask-Socketio
gevent
gevent-websocket
gevent_inotifyx
greenlet
gunicorn
imagehash
inotify
lxml
mistletoe
matplotlib
owoify-py
Pillow
pyotp
pywebpush
qrcode
redis
requests
SQLAlchemy
tinycss2
tldextract
user-agents
psycopg2-binary
youtube-dl
yattag
webptools
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
